Move octoprint coordinator to own file
